# cleanup: don't do stale check when passing --prune.