@fxcoudert 2020-12-31
With that merged, the timings for Intel Big Sur runners have gone from:
- cleanup-before 2m50s 
- cleanup-after 3m0s
to
- cleanup-before 45s
- cleanup-after 1m20s

I'd have to check what is still getting so long. Even with thousands of files, getting `mtime` and `ctime` should not take an entire minute. For reference, even on the ARM runner that has the most cached files, it's taking 5 seconds in the worst case.