@MikeMcQuaid 2020-12-31
> I'd have to check what is still taking so long. Even with thousands of files, getting `mtime` and `ctime` should not take an entire minute. For reference, even on the ARM runner that has the most cached files, it's taking 5 seconds in the worst case.

I don't think the ARM runner is a good comparison for various reasons:
- it's a new architecture that's been custom-designed to run macOS
- it's not running in a VM
- it's using SSD storage

`find -delete` could be an acceptable replacement to put in `brew test-bot` itself instead of `brew cleanup --prune`.